Move dpr change detection to Timeline
